Revert atomic directory structure
