+++ src/components/claims/claims-list/components/LargeClaimsList.tsx
@@ -1,19 +1,22 @@
 import styled from '@emotion/styled'
 import { ClaimState } from 'api/generated/graphql'
-import { ClaimTypeTableCell } from 'components/molecules/ClaimTypeTableCell'
-import { DateTimeTableCell } from 'components/molecules/DateTimeTableCell'
+import { MemberInfoTableCell } from 'components/molecules/MemberInfoTableCell'
 import { parseISO } from 'date-fns'
-import { useGetMemberClaims } from 'graphql/use-get-member-claims'
+import formatDate from 'date-fns/format'
+import { useListClaims } from 'graphql/use-list-claims'
+import { LoadingMessage } from 'hedvig-ui/animations/standalone-message'
 import {
-  LoadingMessage,
-  StandaloneMessage,
-} from 'hedvig-ui/animations/standalone-message'
-import { Table, TableColumn, TableHeader, TableRow } from 'hedvig-ui/table'
+  Table,
+  TableColumn,
+  TableHeader,
+  TablePageSelect,
+  TableRow,
+} from 'hedvig-ui/table'
 import { Monetary, Placeholder } from 'hedvig-ui/typography'
-import React from 'react'
+import React, { useEffect } from 'react'
 import { useHistory } from 'react-router'
 import { useVerticalKeyboardNavigation } from 'utils/keyboard-actions'
-import { convertEnumToTitle } from 'utils/text'
+import { convertEnumToTitle, splitOnUpperCase } from 'utils/text'
 
 const ClaimStateBadge = styled.span<{ state: ClaimState }>`
   display: inline-block;
@@ -27,18 +30,32 @@ const ClaimStateBadge = styled.span<{ state: ClaimState }>`
   border-radius: 8px;
 `
 
-export const MemberClaimsList: React.FC<{ memberId: string }> = ({
-  memberId,
-}) => {
-  const history = useHistory()
-  const [memberClaims, { loading }] = useGetMemberClaims(memberId)
+const TableColumnSubtext = styled.span`
+  font-size: 0.8em;
+  color: ${({ theme }) => theme.semiStrongForeground};
+`
 
-  const claims = memberClaims ?? []
+const FlexVertically = styled.div`
+  display: flex;
+  flex-direction: column;
+`
+
+export const LargeClaimsList: React.FC<{ page: number }> = ({ page }) => {
+  const history = useHistory()
+  const [
+    { claims, page: currentPage, totalPages },
+    listClaims,
+  ] = useListClaims()
+
+  useEffect(() => {
+    listClaims({ page: page - 1 ?? 0 })
+  }, [page])
 
   const [currentKeyboardNavigationStep] = useVerticalKeyboardNavigation({
     maxStep: claims.length - 1,
     onPerformNavigation: (index) => {
       const claimId = claims[index].id
+      const memberId = claims[index].member?.memberId
 
       if (!claimId || !memberId) {
         return
@@ -48,28 +65,30 @@ export const MemberClaimsList: React.FC<{ memberId: string }> = ({
     },
   })
 
-  if (loading) {
-    return <LoadingMessage paddingTop={'10vh'} />
-  }
-
-  if (!claims || claims.length === 0) {
-    return (
-      <StandaloneMessage paddingTop={'10vh'}>
-        No claims for member
-      </StandaloneMessage>
-    )
+  if (!claims) {
+    return <LoadingMessage paddingTop={'25vh'} />
   }
 
   return (
     <>
       <Table>
         <TableRow>
+          <TableHeader>Member</TableHeader>
           <TableHeader>Date Registered</TableHeader>
           <TableHeader>Claim Type</TableHeader>
           <TableHeader>Claim State</TableHeader>
           <TableHeader>Claim Reserves</TableHeader>
         </TableRow>
         {claims.map((claim, index) => {
+          const registrationDateString = formatDate(
+            parseISO(claim.registrationDate),
+            'dd MMMM, yyyy',
+          )
+          const registrationDateTimeString = formatDate(
+            parseISO(claim.registrationDate),
+            'HH:mm',
+          )
+
           return (
             <TableRow
               key={claim.id}
@@ -81,11 +100,24 @@ export const MemberClaimsList: React.FC<{ memberId: string }> = ({
               }
             >
               <TableColumn>
-                <DateTimeTableCell date={parseISO(claim.registrationDate)} />
+                <MemberInfoTableCell member={claim.member} />
               </TableColumn>
 
               <TableColumn>
-                <ClaimTypeTableCell type={claim.type} />
+                <FlexVertically>
+                  {registrationDateString}
+                  <TableColumnSubtext>
+                    {registrationDateTimeString}
+                  </TableColumnSubtext>
+                </FlexVertically>
+              </TableColumn>
+
+              <TableColumn>
+                {claim.type?.__typename ? (
+                  splitOnUpperCase(claim.type.__typename.toString())
+                ) : (
+                  <Placeholder>Not specified</Placeholder>
+                )}
               </TableColumn>
 
               <TableColumn>
@@ -105,6 +137,11 @@ export const MemberClaimsList: React.FC<{ memberId: string }> = ({
           )
         })}
       </Table>
+      <TablePageSelect
+        currentPage={currentPage}
+        totalPages={totalPages}
+        onSelect={(newPage) => history.push(`/claims/list/${newPage}`)}
+      />
     </>
   )
 }
