+++ src/components/members-search/index.tsx
@@ -1,3 +1,4 @@
+import { MembersList } from 'components/members-search/components/MembersList'
 import { SearchForm } from 'components/members-search/components/SearchForm'
 import {
   ExtraInstruction,
@@ -6,10 +7,9 @@ import {
   MemberSuggestionsWrapper,
   NoMembers,
 } from 'components/members-search/styles'
-import { TablePageSelect } from 'components/molecules/TablePageSelect'
-import { MembersList } from 'components/organisms/MembersList'
 import { useMemberSearch } from 'graphql/use-member-search'
 import { FadeIn } from 'hedvig-ui/animations/fade-in'
+import { TablePageSelect } from 'hedvig-ui/table'
 import { MainHeadline } from 'hedvig-ui/typography'
 import React, { useRef } from 'react'
 import { findDOMNode } from 'react-dom'
